[jsk_tools] move ws_doctor.py's function to sanity_lib.py
