+++ jsk_tools/src/jsk_tools/sanity_lib.py
@@ -3,6 +3,7 @@ import rospy
 import rosnode
 import rostopic
 import rosgraph
+import rospkg
 import os
 import subprocess
 from threading import Lock
@@ -26,6 +27,23 @@ def indexMessage(msg):
     print Fore.LIGHTCYAN_EX + "  == %s ==" % (msg) + Fore.RESET
 
 from sensor_msgs.msg import Image, JointState, Imu
+
+def colored(string, color):
+    colors = {
+        'clear': '\033[0m',
+        'black': '\033[30m',
+        'red': '\033[31m',
+        'green': '\033[32m',
+        'yellow': '\033[33m',
+        'blue': '\033[34m',
+        'purple': '\033[35m',
+        'cyan': '\033[36m',
+        'white': '\033[37m'
+        }
+    if colors.has_key(color):
+        return colors[color] + string + colors['clear']
+    else:
+        return string
 
 class TopicPublishedChecker():
     is_topic_published = False
@@ -337,3 +355,160 @@ def checkBlackListDaemon(daemon_names, kill=False):
 
         else:
             okMessage("No BAD Daemon " + daemon_name + " found.")
+
+class ROSMasterChecker():
+    def __init__(self):
+        self.rosmater_pid = None
+        self.cur_tracking_pid = None
+
+    def getProcessNameFromPID(self, pid):
+        return subprocess.check_output("ps -p  %s -o comm= " % (str(pid)), shell=True).strip("\n")
+
+    def getParentPID(self, pid):
+        return subprocess.check_output("ps -p  " + str(pid) + " -oppid= ", shell=True).strip("\n")
+
+    def getMasterPID(self):
+        self.rosmaster_pid = subprocess.check_output("sudo ps aux | grep rosmaster | cut -f 3 -d ' ' |  sed -n '1p'", shell=True).strip("\n")
+
+    def checkIsFromRoscore(self):
+        self.getMasterPID()
+        self.cur_tracking_pid = self.rosmaster_pid
+        cur_tracking_pname = self.getProcessNameFromPID(self.cur_tracking_pid)
+        while not cur_tracking_pname in [ "bash", "gnome-terminal"]:
+            cur_tracking_pname = self.getProcessNameFromPID(self.cur_tracking_pid)
+            if cur_tracking_pname == "roslaunch":
+                errorMessage("This rosmaster is generated by ROSLAUNCH !!")
+                return False
+            self.cur_tracking_pid = self.getParentPID(self.cur_tracking_pid)
+        else:
+            okMessage("This rosmaster is generated by roscore")
+            return True
+
+def checkROSCoreROSMaster():
+    indexMessage("Check ROSmaster came from roscore")
+    rmc = ROSMasterChecker()
+    rmc.checkIsFromRoscore()
+
+
+def colored(string, color):
+    colors = {
+        'clear': '\033[0m',
+        'black': '\033[30m',
+        'red': '\033[31m',
+        'green': '\033[32m',
+        'yellow': '\033[33m',
+        'blue': '\033[34m',
+        'purple': '\033[35m',
+        'cyan': '\033[36m',
+        'white': '\033[37m'
+        }
+    if colors.has_key(color):
+        return colors[color] + string + colors['clear']
+    else:
+        return string
+
+from operator import add
+
+def isROSWS():
+    if os.environ.has_key("ROS_WORKSPACE"):
+        ROS_WORKSPACE = os.environ["ROS_WORKSPACE"]
+        return (ROS_WORKSPACE and 
+                os.path.exists(os.path.join(ROS_WORKSPACE, ".rosinstall")))
+    else:
+        return False
+
+def splitPathEnv(env_name):
+    return env_name.split(":")
+
+def estimateROSPackagePath(path):
+    if path.endswith("devel"):
+        return [path, os.path.join(path, "..", "src")]
+    else:
+        return [os.path.join(path, "share"), os.path.join(path, "stacks")]
+
+def checkROSPackagePath():
+    ROS_PACKAGE_PATH = splitPathEnv(os.environ["ROS_PACKAGE_PATH"])
+    CMAKE_PREFIX_PATH = splitPathEnv(os.environ["CMAKE_PREFIX_PATH"])
+    estimated_ros_package_path = [os.path.abspath(p) 
+                                  for p in 
+                                  reduce(add, [estimateROSPackagePath(c) 
+                                               for c in CMAKE_PREFIX_PATH])]
+    dubious_paths = []
+    for p in ROS_PACKAGE_PATH:
+        normalized_path = os.path.abspath(p)
+        if not normalized_path in estimated_ros_package_path:
+            if isROSWS():
+                ROS_WORKSPACE = os.path.abspath(os.environ["ROS_WORKSPACE"])
+                if not p.startswith(ROS_WORKSPACE):
+                    dubious_paths.append(p)
+    print colored("[ROS_PACKAGE_PATH]", "green")
+    if dubious_paths:
+        print "  ", colored("these path might be malformed: ", "red")
+        for p in dubious_paths:
+            print "    ", colored(p, "red")
+    else:
+        print "  ", colored("ROS_PACKAGE_PATH seems to be OK", "cyan")
+
+
+def checkGitRepoDiff(git_path):
+    os.chdir(git_path)
+    output = subprocess.check_output(["git", "status", "--short"])
+    modified_files = []
+    if output:
+        for l in output.split("\n"):
+            if l.startswith(" M"):
+                modified_files.append(l)
+    if modified_files:
+        print colored("  %d files are modified" % (len(modified_files)), "red")
+        for f in modified_files:
+            print "    ", colored(f, "red")
+    else:
+        print okMessage("  No modified files")
+            
+def checkGitBranch(git_path):
+    os.chdir(git_path)
+    current_branch = subprocess.check_output(["git", "rev-parse", "--abbrev-ref",
+                                   "HEAD"]).strip()
+    try:
+        with open(os.devnull, "w") as devnull:
+            current_tracking_branch = subprocess.check_output(["git", "rev-parse",
+                                                    "--abbrev-ref",
+                                                    "--symbolic-full-name", "@{u}"],
+                                                    stderr=devnull).strip()
+    except subprocess.CalledProcessError:
+        current_tracking_branch = None
+    remote_origin_head = subprocess.check_output(["git", "rev-parse", "--abbrev-ref",
+                                       "origin/HEAD"]).strip()
+    if not current_tracking_branch:
+        print "    ", colored("no tracking branch", "red")
+    elif current_tracking_branch != remote_origin_head:
+        print "    ", colored("the branch(%s) may not sync with %s" 
+                              % (current_branch,
+                              remote_origin_head), "red")
+    else:
+        print okMessage("  No Branch Problem")
+
+        
+def checkGitRepo(git_path):
+    print colored("[checking %s]" % git_path, "green")
+    checkGitRepoDiff(git_path)
+    checkGitBranch(git_path)
+
+def checkGitRepoWithRosPack(rospackage_name):
+    checkGitRepo(rospkg.RosPack().get_path(rospackage_name))
+
+def checkWorkspace():
+    workspace = None
+    if isROSWS():
+        workspaces = [os.environ["ROS_WORKSPACE"]]
+    else:
+        workspaces = [os.path.abspath(os.path.join(p, "..", "src"))
+                      for p in splitPathEnv(os.environ["CMAKE_PREFIX_PATH"])]
+        workspaces.reverse()
+    git_repos = []
+    for workspace in workspaces:
+        for root, dirs, files in os.walk(workspace):
+            if ".git" in dirs:                  
+                if not [repo for repo in git_repos if root.startswith(repo)]: #ignore subdirs
+                    git_repos.append(root)
+                    checkGitRepo(root)
